Finished Delete function

+++ public/js/delete.js
@@ -1,14 +1,20 @@
-const deleteHandler = async () => {
-    const response = await fetch("/api/users/delete", {
-        method: "DELETE",
-        headers: { "Content-Type": "application/json" },
+const deleteHandler = async (e) => {
+
+  const element = e.target;
+  if (element.matches("button")) {
+    const id = element.dataset.id;
+      
+    const response = await fetch(`/${id}`, {
+      method: "DELETE",
+      headers: { "Content-Type": "application/json" },
     });
-    if(response.ok){
-        document.location.delete();
+  
+    if (response.ok) {
+      return location.reload();
     }
-    return
+  }
 };
 
 document
-    .querySelector("#deleteCharacter")
-    .addEventListener("click", deleteHandler);
+  .querySelector("#deleteCharacter")
+  .addEventListener("click", deleteHandler);
